Made MAC address of first NIC configurable
The second argument to the create_nova_node.sh script can be a MAC address which will 
be used to set the MAC of the first NIC. If nothing is supplied the address is auto
generated.

Using this technique we can create predictable MACs and therefore predictable 
automated installs of our Crowbar and Openstack environment.